@@ -3,51 +3,51 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "pretty-quick": "pretty-quick",
+    "dev": "vite",
+    "build": "run-p type-check build-only",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit -p ./tsconfig.app.json --composite false",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
     "prepare": "husky install"
   },
   "prettier": "@bokub/prettier-config",
   "dependencies": {
     "@mapbox/polyline": "^1.1.1",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.0.0-beta2",
+    "axios": "^0.24.0",
+    "bootstrap": "^5.3.0",
     "cookie": "^0.4.1",
-    "core-js": "^3.6.5",
     "debounce": "^1.2.0",
     "js-cookie": "^2.2.1",
     "leaflet": "^1.7.1",
     "leaflet.locatecontrol": "^0.72.2",
-    "register-service-worker": "^1.7.1",
     "strava-v3": "^2.0.8",
-    "vue": "^3.0.6"
+    "vue": "^3.3.4"
   },
   "devDependencies": {
     "@bokub/prettier-config": "^2.1.0",
+    "@rushstack/eslint-patch": "^1.2.0",
+    "@tsconfig/node18": "^2.0.1",
     "@types/debounce": "^1.2.0",
     "@types/js-cookie": "^2.2.6",
     "@types/leaflet": "^1.5.23",
     "@types/mapbox__polyline": "^1.0.2",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-pwa": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^5.0.2",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^7.0.0-0",
-    "husky": "^6.0.0",
-    "node-sass": "^5.0.0",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0",
-    "sass-loader": "^10.1.1",
-    "typescript": "~3.9.3"
+    "@types/node": "^18.16.17",
+    "@vercel/node": "^2.15.4",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/eslint-config-prettier": "^7.1.0",
+    "@vue/eslint-config-typescript": "^11.0.3",
+    "@vue/tsconfig": "^0.4.0",
+    "eslint": "^8.39.0",
+    "eslint-plugin-vue": "^9.11.0",
+    "husky": "^8.0.3",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.8.8",
+    "pretty-quick": "^3.1.3",
+    "sass": "^1.63.6",
+    "typescript": "~5.0.4",
+    "vite": "^4.3.9",
+    "vite-plugin-pwa": "^0.16.4",
+    "vue-tsc": "^1.6.5"
   }
 }
